% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.attachment.R
\name{api.add_attachment}
\alias{api.add_attachment}
\title{Add an attachment to the specified model.}
\usage{
api.add_attachment(model_name, attachment_file)
}
\arguments{
\item{model_name}{The name of the model}

\item{attachment_file}{The path to the file.}
}
\value{
True, if successful.
}
\description{
Add an attachment to the specified model.
}
