% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.model.R
\name{api.add_model}
\alias{api.add_model}
\title{Add a model to Model Manage.}
\usage{
api.add_model(model_name, model_content, model_type = "R")
}
\arguments{
\item{model_name}{The name of the model}

\item{model_content}{The contents of the model}

\item{model_type}{The language the model is written in (default: R)}
}
\value{
True, if successful.
}
\description{
Add a model to Model Manage.
}
