% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.schema.R
\name{api.add_schema}
\alias{api.add_schema}
\title{Add a schema to Model Manage.}
\usage{
api.add_schema(schema_name, schema_content)
}
\arguments{
\item{schema_name}{The name of the schema in Model Manage}

\item{schema_content}{The Avro schema (a JSON string)}
}
\value{
True if successful.
}
\description{
Add a schema to Model Manage.
}
