% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.stream.R
\name{api.add_stream}
\alias{api.add_stream}
\title{Add the named stream to Model Manage.}
\usage{
api.add_stream(stream_name, stream_content)
}
\arguments{
\item{stream_name}{A name for the stream.}

\item{stream_content}{The stream descriptor.}
}
\value{
True if successful.
}
\description{
Add the named stream to Model Manage.
}
