% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.job.R
\name{api.deploy_model}
\alias{api.deploy_model}
\title{Deploys the named model.}
\usage{
api.deploy_model(model_content, model_name, ctype, attachments = list(),
  container = NULL)
}
\arguments{
\item{model_content}{The contents of the model.}

\item{model_name}{A name for the model.}

\item{ctype}{The content-type disposition for the model.}

\item{container}{The name of the container to use (optional)}
}
\value{
True, if successful.
}
\description{
Deploys the named model.
}
