% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.job.R
\name{api.deploy_output_stream}
\alias{api.deploy_output_stream}
\title{Deploys the named stream to the engine (output).}
\usage{
api.deploy_output_stream(stream_content, stream_name, container = NULL)
}
\arguments{
\item{stream_content}{A stream descriptor string.}

\item{stream_name}{A name for this stream.}

\item{container}{The name of the container to use (optional)}
}
\value{
True, if successful.
}
\description{
Deploys the named stream to the engine (output).
}
