% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.attachment.R
\name{api.get_attachment}
\alias{api.get_attachment}
\title{Retrieve an attachment from a model, and save it to a file.}
\usage{
api.get_attachment(model_name, attachment_name, attachment_path = "")
}
\arguments{
\item{model_name}{The name of the model.}

\item{attachment_name}{The name of the attachment.}

\item{attachment_path}{The path to save the attachment to (optional;
defaults to current working directory)}
}
\value{
True, if successful.
}
\description{
Retrieve an attachment from a model, and save it to a file.
}
