% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.model.R
\name{api.get_model}
\alias{api.get_model}
\title{Retrieve a model's contents from Model Manage.}
\usage{
api.get_model(model_name, include_ctype = FALSE)
}
\arguments{
\item{model_name}{The name of the model to retrieve}

\item{include_ctype}{Whether to return the content-type as well (default: False)}
}
\value{
The contents of the model, and, optionally, its content-type header.
}
\description{
Retrieve a model's contents from Model Manage.
}
