% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.schema.R
\name{api.get_schema}
\alias{api.get_schema}
\title{Retrieve a schema from Model Manage.}
\usage{
api.get_schema(schema_name)
}
\arguments{
\item{schema_name}{The name of the schema in Model Manage.}
}
\value{
The contents of the named schema (a JSON string)
}
\description{
Retrieve a schema from Model Manage.
}
