% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.stream.R
\name{api.get_stream}
\alias{api.get_stream}
\title{Retrieve the stream descriptor for the named stream.}
\usage{
api.get_stream(stream_name)
}
\arguments{
\item{stream_name}{The name of the stream to retrieve.}
}
\value{
A string with the stream descriptor for the named stream.
}
\description{
Retrieve the stream descriptor for the named stream.
}
