% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.job.R
\name{api.job_input}
\alias{api.job_input}
\title{Send inputs to the engine for scoring, and return the results.}
\usage{
api.job_input(input_data, container = NULL)
}
\arguments{
\item{input_data}{The data to send to the engine for scoring (an array of JSON strings)}

\item{container}{The name of the container to use (optional)}
}
\value{
A list of output records (JSON strings).
}
\description{
Send inputs to the engine for scoring, and return the results.
}
