% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.job.R
\name{api.job_status}
\alias{api.job_status}
\title{Retrieve the status of the currently running job on the specified engine.}
\usage{
api.job_status(container = NULL)
}
\arguments{
\item{container}{The name of the container to use (optional)}
}
\value{
A JSON object whose top-level fields are 'jets', 'model', 'input',
        and 'output'
}
\description{
Retrieve the status of the currently running job on the specified engine.
}
