% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.attachment.R
\name{api.list_attachments}
\alias{api.list_attachments}
\title{List the names of all the attachments associated with the given model.}
\usage{
api.list_attachments(model_name)
}
\arguments{
\item{model_name}{The name of the model.}
}
\value{
A list of all the attachments.
}
\description{
List the names of all the attachments associated with the given model.
}
