% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.model.R
\name{api.list_models}
\alias{api.list_models}
\title{List all of the models in Model Manage, by name.}
\usage{
api.list_models()
}
\value{
A list of the names of the models in Model Manage.
}
\description{
List all of the models in Model Manage, by name.
}
