% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.schema.R
\name{api.list_schemata}
\alias{api.list_schemata}
\title{List all of the schemata in Model Manage, by name.}
\usage{
api.list_schemata()
}
\value{
A list of the names of every schema in Model Manage.
}
\description{
List all of the schemata in Model Manage, by name.
}
