% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.attachment.R
\name{api.remove_attachment}
\alias{api.remove_attachment}
\title{Remove the named attachment from the specified model.}
\usage{
api.remove_attachment(model_name, attachment_name)
}
\arguments{
\item{model_name}{The name of the model}

\item{attachment_name}{The name of the attachment}
}
\value{
True, if successful.
}
\description{
Remove the named attachment from the specified model.
}
