% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.model.R
\name{api.remove_model}
\alias{api.remove_model}
\title{Remove the named model from Model Manage.}
\usage{
api.remove_model(model_name)
}
\arguments{
\item{model_name}{The name of the model to remove.}
}
\value{
True, if successful
}
\description{
Remove the named model from Model Manage.
}
