% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.schema.R
\name{api.remove_schema}
\alias{api.remove_schema}
\title{Remove a schema from Model Manage.}
\usage{
api.remove_schema(schema_name)
}
\arguments{
\item{schema_name}{The name of the schema to remove.}
}
\value{
True, if successful.
}
\description{
Remove a schema from Model Manage.
}
