% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.stream.R
\name{api.remove_stream}
\alias{api.remove_stream}
\title{Remove the named stream from Model Manage.}
\usage{
api.remove_stream(stream_name)
}
\arguments{
\item{stream_name}{The name of the stream to remove.}
}
\value{
True, if successful.
}
\description{
Remove the named stream from Model Manage.
}
