% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.job.R
\name{api.run_job}
\alias{api.run_job}
\title{Runs the named model using the specified input and output streams.}
\usage{
api.run_job(model, input_stream, output_stream, container = NULL)
}
\arguments{
\item{model}{The name of the model in Model Manage.}

\item{input_stream}{The name of the input stream descriptor.}

\item{output_stream}{The name of the output stream descriptor.}

\item{container}{The name of the container to use (optional)}
}
\value{
True, if successful.
}
\description{
Runs the named model using the specified input and output streams.
}
