% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.job.R
\name{api.stop_job}
\alias{api.stop_job}
\title{Stop the current job.}
\usage{
api.stop_job(container = NULL)
}
\arguments{
\item{container}{The name of the container to use (optional)}
}
\value{
True if successful.
}
\description{
Stop the current job.
}
