% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.datatype.R
\name{avroTypeToJsonNode}
\alias{avroTypeToJsonNode}
\title{Write an AvroType object to a JSON object.}
\usage{
avroTypeToJsonNode(schema, toString = TRUE)
}
\arguments{
\item{schema}{The AvroType object.}

\item{toString}{Should the result be a string? (Default: TRUE)}
}
\value{
A JSON object.
}
\description{
Write an AvroType object to a JSON object.
}
