% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{checkData}
\alias{checkData}
\title{Check whether the given data matches the given schema.}
\usage{
checkData(data, avroType)
}
\arguments{
\item{data}{The datum to test.}

\item{avroType}{An AvroType object.}
}
\value{
TRUE if data satisfies avroType, and FALSE otherwise.
}
\description{
Check whether the given data matches the given schema.
}
