% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.connect.R
\docType{class}
\name{Connect-class}
\alias{Connect-class}
\alias{Connect}
\title{Connect}
\description{
Reference class for Fastscore Connect instance
}
\section{Fields}{

\describe{
\item{\code{proxy_prefix}}{proxy string}

\item{\code{insecure}}{indicator for insecure curl option}
}}


