% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\docType{data}
\name{Connect}
\alias{Connect}
\title{Create an R6 "connect" class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Connect
}
\value{
An R6 class object
}
\description{
\code{connect} creates an R6 class with methods for connect operations;
it wraps \code{swagger::ConnectApi}
}
\section{Fields}{

\describe{
\item{\code{$path}}{Stores url path of the request.}

\item{\code{$apiClient}}{Handles the client-server communication.}

\item{\code{$userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

\item{\code{$active_sensor_install()}}{install active sensor}

\item{\code{$active_sensor_list()}}

\item{\code{$active_sensor_points()}}

\item{\code{$active_sensor_uninstall()}}

\item{\code{$pneumo_get()}}
}
}

\keyword{datasets}
