% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.connect.R
\name{Connect_configure}
\alias{Connect_configure}
\title{Set Fastscore configuration}
\arguments{
\item{config}{configuration description yaml}
}
\value{
if configuration successfully set
}
\description{
Set Fastscore configuration
}
