% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.connect.R
\name{Connect_get}
\alias{Connect_get}
\title{Find a Fastscore instance by name}
\arguments{
\item{name}{instance name}
}
\value{
an sintance object of the specified name
}
\description{
Find a Fastscore instance by name
}
