% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.connect.R
\name{Connect_lookup}
\alias{Connect_lookup}
\title{Find a Fastscore instance by type}
\arguments{
\item{sname}{instance type name}
}
\value{
an sintance object of the specified type
}
\description{
Find a Fastscore instance by type
}
