% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\docType{data}
\name{Engine}
\alias{Engine}
\title{FastScore Engine operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Engine
}
\description{
fastscoRe::Engine
}
\details{
FastScore Engine (proxy) API Client

An R6 class generator for Engine functions.
Inherits methods and fields from class
generator \code{swaggerv1::EngineApi}.
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

\item{\code{$health_get()}}{}

\item{\code{$input_stream_set()}}{}

\item{\code{$job_delete()}}{}

\item{\code{$job_io_input()}}{}

\item{\code{$job_io_input0()}}{}

\item{\code{$job_io_output()}}{}

\item{\code{$job_io_output1()}}{}

\item{\code{$job_state_restore()}}{}

\item{\code{$job_status()}}{}

\item{\code{$model_load()}}{}

\item{\code{$model_unload()}}{}

\item{\code{$output_stream_set()}}{}

\item{\code{$stream_attach()}}{}
}
}

\keyword{datasets}
