% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.engine.R
\name{Engine_input_set}
\alias{Engine_input_set}
\title{Attaching input stream to an engine}
\arguments{
\item{slot}{attaching slot, use even numbers}

\item{stream}{stream instance to be attached}
}
\value{
if stream is successfully attached
}
\description{
Attaching input stream to an engine
}
