% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.engine.R
\name{Engine_load_model}
\alias{Engine_load_model}
\title{Load a model to an engine}
\arguments{
\item{model}{model instance to be attached}

\item{force_inline}{if bypass preferred list, default = FALSE}
}
\value{
if model is successfully loaded
}
\description{
Load a model to an engine
}
