% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.engine.R
\name{Engine_output_set}
\alias{Engine_output_set}
\title{Attaching output stream to an engine}
\arguments{
\item{slot}{attaching slot, use odd numbers}

\item{stream}{stream instance to be attached}
}
\value{
if stream is successfully attached
}
\description{
Attaching output stream to an engine
}
