% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.engine.R
\name{Engine_score}
\alias{Engine_score}
\title{Run an input to an engine with loaded model, and get output}
\arguments{
\item{data}{input data, which is passed in as a json string}

\item{encode}{default FALSE. TODO: implement auto-encoding according to model schema}
}
\value{
model output
}
\description{
Run an input to an engine with loaded model, and get output
}
