% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.engine.R
\name{Engine_unload_model}
\alias{Engine_unload_model}
\title{Unload all models from an engine, i.e. reset the engine}
\value{
if engine is successfully unloaded
}
\description{
Unload all models from an engine, i.e. reset the engine
}
