% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\docType{data}
\name{FastScoreError}
\alias{FastScoreError}
\title{A FastScore condition (error, warning, message). SDK functions throw only FastScoreError conditions. An SDK function either succeeds or throws a condition. The return value of a SDK function is always valid.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FastScoreError
}
\description{
A FastScore condition (error, warning, message). SDK functions throw only FastScoreError conditions. An SDK function either succeeds or throws a condition. The return value of a SDK function is always valid.
}
\keyword{datasets}
