% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{from_json}
\alias{from_json}
\title{Deserialize a JSON string into an R object.}
\usage{
from_json(datum, schema)
}
\arguments{
\item{datum}{The JSON string to deserialize.}

\item{schema}{The Avro schema to use for deserialization (a string).}
}
\value{
An object.
}
\description{
Deserialize a JSON string into an R object.
}
\examples{
mylist <- as.list(c('{"int":1}', '{"int":2}', '{"int":3}'))
lapply(mylist, from_json, schema='["null", "int"]')
}
