% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{jsonDecoder}
\alias{jsonDecoder}
\title{Decode a JSON object as a given fastscore.datatype.AvroType.}
\usage{
jsonDecoder(avroType, value, fromString = TRUE)
}
\arguments{
\item{avroType}{The AvroType of this object}

\item{value}{A JSON object}

\item{fromString}{If TRUE, value is a JSON string. (Default: TRUE)}
}
\value{
An R object
}
\description{
Decode a JSON object as a given fastscore.datatype.AvroType.
}
