% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.datatype.R
\name{jsonEncoder}
\alias{jsonEncoder}
\title{Encode an object as JSON, given fastscore.datatype.AvroType.}
\usage{
jsonEncoder(avroType, value, tagged = TRUE, toString = TRUE)
}
\arguments{
\item{avroType}{The type of this object.}

\item{value}{The contents to encode.}

\item{tagged}{If true, represent unions as "{tag:value}"; if false, represent them as value.}

\item{toString}{If true, encode the result as a JSON string (default: TRUE).}
}
\value{
a JSON object.
}
\description{
Encode an object as JSON, given fastscore.datatype.AvroType.
}
