% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.datatype.R
\name{jsonNodeToAvroType}
\alias{jsonNodeToAvroType}
\title{Convert a JSON Avro schema to an AvroType.}
\usage{
jsonNodeToAvroType(x, fromString = TRUE)
}
\arguments{
\item{x}{The JSON Avro schema}

\item{fromString}{Should x be first deserialized from a string? (Default: TRUE)}
}
\description{
Convert a JSON Avro schema to an AvroType.
}
