% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\alias{Model}
\title{Model}
\description{
A class that represents a FastScore Model
}
\section{Fields}{

\describe{
\item{\code{model_manage}}{the modelmanage it belongs to}

\item{\code{name}}{model name}

\item{\code{mtype}}{model type}

\item{\code{source}}{model source code}
}}


