% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{Model}
\alias{Model}
\title{FastScore Model Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Model
}
\description{
fastscoRe::Model$new()
}
\details{
FastScore Model Class

An R6 class generator for
FastScore model object representations.
}
\section{Fields}{

\describe{
\item{\code{name}}{A name for the model}

\item{\code{mtype}}{model type; e.g. 'R'}

\item{\code{source}}{the model code}

\item{\code{model_manage}}{the Model Manage instance the model belongs to}
}}

\section{Methods}{

\describe{
\item{\code{$attachment_list}}{}
\item{\code{$attachment_get()}}{}
\item{\code{$attachment_download()}}{}
\item{\code{$attachment_delete()}}{}
}
}

\keyword{datasets}
