% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Model_attachment_delete}
\alias{Model_attachment_delete}
\title{Delete an attachments from this model}
\arguments{
\item{name}{attachment name}
}
\value{
if an attachment is successfully removed
}
\description{
Delete an attachments from this model
}
