% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelManage.R
\docType{data}
\name{ModelManage}
\alias{ModelManage}
\title{RC implementation of Queue}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ModelManage
}
\description{
A simple example of what is possible using roxygen 4.1.1
}
\section{Fields}{

\describe{
\item{\code{queue}}{a list of items in the queue}
}}

\examples{
(qrc <- QueueRC(5, 6, "foo"))
qrc$add("add something")
qrc$add("and something else")
qrc$remove()
}
\keyword{datasets}
