% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.model_manage.R
\name{ModelManage_model_add_attachment}
\alias{ModelManage_model_add_attachment}
\title{Add an local attachment to a model}
\arguments{
\item{model_name}{model name}

\item{attachment_name}{attachment name, including file extensions for attachment type}

\item{attachment_path}{path of the attachment file}
}
\value{
if an attachment is successfully loaded
}
\description{
Add an local attachment to a model
}
