% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.model_manage.R
\name{ModelManage_model_load_from_file}
\alias{ModelManage_model_load_from_file}
\title{Load a model from local file}
\arguments{
\item{name}{model name, including file extension for model language}

\item{path}{path of the model file}
}
\value{
if a model is successfully loaded
}
\description{
Load a model from local file
}
