% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.model_manage.R
\name{ModelManage_schema_get}
\alias{ModelManage_schema_get}
\title{Get a schema object}
\arguments{
\item{name}{schema name}
}
\value{
corresponding schema object
}
\description{
Get a schema object
}
