% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.model_manage.R
\name{ModelManage_schema_load_from_file}
\alias{ModelManage_schema_load_from_file}
\title{Load a schema from local file}
\arguments{
\item{name}{schema name, NOT including file extension, as only supporting .avsc, and to match up with model script smart comment}

\item{path}{path of the schema file}
}
\value{
if a schema is successfully loaded
}
\description{
Load a schema from local file
}
