% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.model_manage.R
\name{ModelManage_sensor_get}
\alias{ModelManage_sensor_get}
\title{Get a sensor object}
\arguments{
\item{name}{sensor name}
}
\value{
corresponding sensor object
}
\description{
Get a sensor object
}
