% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.model_manage.R
\name{ModelManage_stream_get}
\alias{ModelManage_stream_get}
\title{Get a stream object}
\arguments{
\item{name}{stream name}
}
\value{
corresponding stream object
}
\description{
Get a stream object
}
