% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite.model_manage.R
\name{ModelManage_stream_load_from_file}
\alias{ModelManage_stream_load_from_file}
\title{Load a stream from local file}
\arguments{
\item{name}{stream name, NOT including file extension, as only supporting .json}

\item{path}{path of the stream file}
}
\value{
if a stream is successfully loaded
}
\description{
Load a stream from local file
}
