% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{recordset_to_json}
\alias{recordset_to_json}
\title{Serialize a record set into a list of JSON strings}
\usage{
recordset_to_json(recordset, schema)
}
\arguments{
\item{recordset}{The record set (a Data Frame or Matrix)}

\item{schema}{The Avro schema to use}
}
\description{
Serialize a record set into a list of JSON strings
}
