% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{class}
\name{Schema-class}
\alias{Schema-class}
\alias{Schema}
\title{Schema}
\description{
A class that represents a FastScore Schema
}
\section{Fields}{

\describe{
\item{\code{name}}{schema name}

\item{\code{source}}{schema source code}

\item{\code{model_manage}}{the modelmanage it belongs to}
}}


