% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\docType{class}
\name{Stream-class}
\alias{Stream-class}
\alias{Stream}
\title{Stream}
\description{
A class that represents a FastScore Stream
}
\section{Fields}{

\describe{
\item{\code{name}}{stream name}

\item{\code{desc}}{stream description}

\item{\code{model_manage}}{the modelmanage it belongs to}
}}


