% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\docType{data}
\name{Stream}
\alias{Stream}
\title{FastScore Stream Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Stream
}
\description{
fastscoRe::Stream$new()
}
\details{
FastScore Stream Class

An R6 class generator for
FastScore stream object representations.
}
\section{Fields}{

\describe{
\item{\code{name}}{A name for the stream}

\item{\code{source}}{the stream source code}

\item{\code{model_manage}}{the Model Manage instance, if any, the stream belongs to}
}}

\section{Methods}{

\describe{
\item{\code{$method1()}}{}
\item{\code{$method2()}}{}
}
}

\keyword{datasets}
