% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{to_json}
\alias{to_json}
\title{Convert the specified datum into JSON strings using the schema}
\usage{
to_json(datum, schema)
}
\arguments{
\item{datum}{An R object to encode.}

\item{schema}{An Avro schema to use to encode (a string).}
}
\value{
A JSON string.
}
\description{
Convert the specified datum into JSON strings using the schema
}
\examples{
mylist <- as.list(c(1, 2, 3, 4))
lapply(mylist, to_json, schema='["null", "int"]')
}
