% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\docType{data}
\name{Instance}
\alias{Instance}
\title{R client for the FastScore API}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Instance
}
\description{
An R6 class to instantiate an R client for the
FastScore API
}
\section{Fields}{

\describe{
\item{\code{basePath}}{the HTTPS scheme URL for the FastScore Microservices
Dashboard being used for the FastScore instance you wish to connect to}
}}

\keyword{datasets}
