% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelManage.R
\docType{data}
\name{ModelManage}
\alias{ModelManage}
\title{ModelManage operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ModelManage
}
\description{
fastscoRe::ModelManage
}
\details{
FastScore Model Manage (proxy) API Client

An R6 class generator for Model Manage functions.
Inherits methods and fields from class
generator \code{swaggerv1::ModelManageApi}.
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\keyword{datasets}
